Transform event content for API endpoint calls
Passes query string parameters to the handler instead of the whole event in the case that an HTTP request with query string params comes through.

This still doesn't quite cut it for API Gateway: although good Lambda calls POST 200 OK, the API Gateway receives it and sends on "502: Malformed Lambda proxy response". See:
  * https://aws.amazon.com/premiumsupport/knowledge-center/malformed-502-api-gateway
  * https://docs.aws.amazon.com/apigateway/latest/developerguide/set-up-lambda-proxy-integrations.html#api-gateway-simple-proxy-for-lambda-output-format

Not sure how to alter the POST response yet, though!